refactor PdJobInScheduler
