+++ src/CraneCtld/LicensesManager.cpp
@@ -17,6 +17,7 @@
  */
 
 #include "LicensesManager.h"
+#include "TaskScheduler.h"
 
 namespace Ctld {
