+++ src/CraneCtld/RpcService/CtldGrpcServer.cpp
@@ -28,6 +28,7 @@
 #include "Lua/LuaJobHandler.h"
 #include "Security/VaultClient.h"
 #include "TaskScheduler.h"
+#include "LicensesManager.h"
 #include "crane/PluginClient.h"
 #include "protos/PublicDefs.pb.h"
